pyproject.toml
[project]
name = "papergraph"
version = "0.1.0"
description = "LangGraph-based scientific paper processing with storage to BigQuery."
authors = [
    {name = "Daniel Nobbe",email = "daniellnobbe@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "langchain (>=0.3.17,<0.4.0)",
    "langchain-mistralai (>=0.2.6,<0.3.0)",
    "langgraph (>=0.2.69,<0.3.0)",
    "pypdf (>=5.2.0,<6.0.0)",
    "mistralai (>=1.5.0,<2.0.0)",
    "langchain-community (>=0.3.16,<0.4.0)",
    "google-cloud-bigquery (>=3.29.0,<4.0.0)",
    "google-auth (>=2.38.0,<3.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^25.1.0"
pylint = "^3.3.4"
ruff = "^0.9.4"
mypy = "^1.14.1"

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
indent-width = 4

[tool.pylint.messages_control]
disable = ["C0116", "C0114", "C0115"]